pyproject.toml
[project]
name = "route-ai"
version = "0.1.0"
description = "PCB Autorouting Autoencoder"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "datasets>=2.14.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
]
